requirements.txt
Flask==2.3.3
google-generativeai==0.6.0
urllib3==1.26.16
python-dotenv